Update name of PropertyFiltering to PropertyContainer and add routes in app
