+++ src/Components/App/App.js
@@ -1,10 +1,11 @@
 import React, { Component } from 'react';
 import '../../reset.scss';
 import './App.scss';
+import { Route } from 'react-router-dom'
 
 import NavBar from '../NavBar/NavBar';
 import SearchProperty from '../SearchProperty/SearchProperty';
-import PropertyFiltering from '../PropertyFiltering/PropertyFiltering';
+import PropertyFiltering from '../PropertyContainer/PropertyContainer';
 import PropertyContainer from '../PropertyContainer/PropertyContainer';
 
 class App extends Component {
@@ -20,9 +21,8 @@ class App extends Component {
     return(
       <main className='App'>
         <NavBar />
-        <SearchProperty />
-        <PropertyFiltering />
-        <PropertyContainer />
+        <Route exact path='/' render= { () => <SearchProperty /> } />
+        <Route exact path='/' render= { () => <PropertyFiltering /> } />
       </main>
     )
   }
