--- src/Components/PropertyFiltering/PropertyFiltering.js
@@ -1,18 +1,0 @@
-import React from 'react';
-import './PropertyFiltering.scss';
-// import Favourites from './Favourites/Favourites';
-// import BookingsCalendar from './BookingsCalendar/BookingsCalendar';
-// import AddProperty from './AddProperty/AddProperty';
-
-
-const PropertyFiltering = () => {
-  return(
-    <section className='property-filtering'>
-      {/* <Favourites />
-      <BookingsCalendar />
-      <AddProperty /> */}
-    </section>
-  )
-}
-
-export default PropertyFiltering;
